# Add new optimized TX BPF implementation for OFDM modes.