--- a/src/filter.c
+++ b/src/filter.c
@@ -19,6 +19,7 @@
 #include <string.h>
 #include <math.h>
 #include <complex.h>
+#include <assert.h>
 
 #include "filter.h"
 #include "filter_coef.h"
@@ -46,6 +47,7 @@ void quisk_filt_cfInit(struct quisk_cfFilter * filter, float * coefs, int taps) 
     // be real or complex.
     filter->dCoefs = coefs;
     filter->cpxCoefs = NULL;
+    filter->realCoefs = NULL;
     filter->cSamples = (complex float *)MALLOC(taps * sizeof(complex float));
     memset(filter->cSamples, 0, taps * sizeof(complex float));
     filter->ptcSamp = filter->cSamples + taps - 1;
@@ -80,6 +82,11 @@ void quisk_filt_destroy(struct quisk_cfFilter * filter) {
     if (filter->cpxCoefs) {
         FREE(filter->cpxCoefs);
         filter->cpxCoefs = NULL;
+    }
+    
+    if (filter->realCoefs) {
+        FREE(filter->realCoefs);
+        filter->realCoefs = NULL;
     }
 }
 
@@ -229,14 +236,19 @@ void quisk_cfTune(struct quisk_cfFilter * filter, float freq) {
     int i;
 
     if ( ! filter->cpxCoefs)
+    {
         filter->cpxCoefs = (complex float *)MALLOC(filter->nTaps * sizeof(complex float));
-
+        filter->realCoefs = (float *)MALLOC(filter->nTaps * sizeof(float));
+        assert(filter->cpxCoefs && filter->realCoefs);
+    }
+    
     tune = 2.0 * M_PI * freq;
     D = (filter->nTaps - 1.0) / 2.0;
 
     for (i = 0; i < filter->nTaps; i++) {
         float tval = tune * (i - D);
         filter->cpxCoefs[i] = cmplx(tval) * filter->dCoefs[i];
+        filter->realCoefs[i] = cabsf(filter->cpxCoefs[i]);
     }
 }
 
@@ -294,12 +306,11 @@ void quisk_ccfFilter(complex float * inSamples, complex float * outSamples, int 
 void quisk_crfFilter(complex float * inSamples, complex float * outSamples, int count, struct quisk_cfFilter * filter) {
     for (int i = 0; i < count; i++) 
     {    
-        float* coeffPtr = (float*)filter->cpxCoefs;
+        float* coeffPtr = filter->realCoefs;
         int numValsInTaps = filter->nTaps * 2;
-        float filterOutAC = 0;
-        float filterOutBD = 0;
+        float filterOut = 0;
         
-        *filter->ptcSamp = inSamples[i];
+        *filter->ptcSamp = cabsf(inSamples[i]);
         
         float* ptSample = (float*)filter->ptcSamp;
         float* endPtSample = (float*)(filter->cSamples + filter->nTaps);
@@ -308,16 +319,16 @@ void quisk_crfFilter(complex float * inSamples, complex float * outSamples, int 
             // Note: The below only calculates the real component of the result,
             // not the complex one. This works because:
             // (a + bi)(c + di) = ac + bci + adi + bdi^2 = ac - bd + (ad + bc)i.
-            filterOutAC += *ptSample++ * *coeffPtr++;
-            filterOutBD -= *ptSample++ * *coeffPtr++;
+            filterOut += *ptSample * *coeffPtr++;
+            ptSample += 2;
             
             if (ptSample == endPtSample)
                 ptSample = (float*)filter->cSamples;
         }
 
-        outSamples[i] = filterOutAC + filterOutBD;
+        outSamples[i] = filterOut;
         
         if (--filter->ptcSamp < filter->cSamples)
-            filter->ptcSamp = filter->cSamples + filter->nTaps - 1;
+            filter->ptcSamp = (filter->cSamples + filter->nTaps - 1);
     }
 }