This PR contains a potential optimized TX BPF implementation that basically ignores the complex component and only operates on real components. As a result, there is a significant improvement in performance on x86_64:

*Original implementation*:

```
HT-TM05:build_osx mooneer$ time src/freedv_tx --clip 1 --txbpf 1 700D hrlongsinewave.raw tx.raw

real	0m17.933s
user	0m17.444s
sys	0m0.095s
HT-TM05:build_osx mooneer$ time src/freedv_tx --clip 1 --txbpf 1 700D hrlongsinewave.raw tx.raw

real	0m17.501s
user	0m17.326s
sys	0m0.092s
HT-TM05:build_osx mooneer$ time src/freedv_tx --clip 1 --txbpf 1 700D hrlongsinewave.raw tx.raw

real	0m18.805s
user	0m18.241s
sys	0m0.138s
HT-TM05:build_osx mooneer$ 
```

*This PR*:

```
HT-TM05:build_osx mooneer$ time src/freedv_tx --clip 1 --txbpf 1 700D hrlongsinewave.raw tx.raw

real	0m13.571s
user	0m13.470s
sys	0m0.091s
HT-TM05:build_osx mooneer$ time src/freedv_tx --clip 1 --txbpf 1 700D hrlongsinewave.raw tx.raw

real	0m13.621s
user	0m13.491s
sys	0m0.092s
HT-TM05:build_osx mooneer$ time src/freedv_tx --clip 1 --txbpf 1 700D hrlongsinewave.raw tx.raw

real	0m13.513s
user	0m13.415s
sys	0m0.090s
HT-TM05:build_osx mooneer$
```

SNR per `freedv_rx` also appears similar to the current implementation:

```
HT-TM05:build_osx mooneer$ src/freedv_rx -v 700D tx.raw rx.raw
frame: 1  demod sync: 0  nin: 1280 demod snr: -5.00 dB  bit errors: 0 clock_offset: 0.000000
frame: 2  demod sync: 0  nin: 370 demod snr: -5.00 dB  bit errors: 0 clock_offset: 0.000000
frame: 3  demod sync: 1  nin: 1280 demod snr: 6.76 dB  bit errors: 0 clock_offset: 0.000000
frame: 4  demod sync: 1  nin: 1280 demod snr: 11.43 dB  bit errors: 0 clock_offset: 0.000000
frame: 5  demod sync: 1  nin: 1280 demod snr: 11.10 dB  bit errors: 0 clock_offset: 0.000000
frame: 6  demod sync: 1  nin: 1280 demod snr: 10.86 dB  bit errors: 0 clock_offset: 0.000000
frame: 7  demod sync: 1  nin: 1280 demod snr: 10.53 dB  bit errors: 0 clock_offset: 0.000000
frame: 8  demod sync: 1  nin: 1280 demod snr: 10.36 dB  bit errors: 0 clock_offset: 0.000000
frame: 9  demod sync: 1  nin: 1280 demod snr: 10.09 dB  bit errors: 0 clock_offset: 0.000000
frame: 10  demod sync: 1  nin: 1280 demod snr: 9.97 dB  bit errors: 0 clock_offset: 0.000000
```