@tmiw 2022-06-01
@drowe67, are the assumptions made here reasonable? I'm also not sure if the 100 pole filter the PR is currently using for 700D is sufficient. From the code:

```
// Alternate version of above BPF generated using the following:
//
//     # taps = 100
//     0- 750 Hz: -60dB attenuation
//  1000-2000 Hz: 0.1dB ripple
//  2250-4000 Hz: -60dB attenuation
//
// This filter provides less steep skirts, which allows us to use fewer poles to
// provide bandpass filtering. As a result, there is a significant performance 
// improvement over the 100 pole Quisk complex filter above.
```

I had also tried the following BPF but there was no performance improvement (and might have had slightly worse runtime, even):

```
// BPF generated from http://t-filter.engineerjs.com/ using the following:
//
//     # taps = 256
//     0- 900 Hz: -60dB attenuation
//  1000-2000 Hz: 0.1dB ripple
//  2100-4000 Hz: -60dB attenuation
//
// Note that the coefficients are reversed. This is because after we add the current x[t] to the
// lookback buffer, we start calculating from x[t - N] and loop back around, making b[0] * x[t] the
// the last multiplication done. Doing it this way enables easier vectorization by the compiler 
// (or via e.g. CMSIS).
//
// Additionally, this filter is used as-is by fir_quick_*. That is, the coefficients are not made
// complex before applying them. This works as the TX BPF for OFDM modes is the last operation executed,
// after which the complex component is thrown out.
```

If the less steep skirts are okay, I can migrate 700E, etc. over too. 👍