@drowe67 2022-06-01
Hi @tmiw - thanks for taking a look at optimising the BPF.  It is quite intensive in terms of CPU.

Some thoughts/questions:
1. On what platforms do we actually have a problem with CPU for this filter?  As per the Flex work - I'm uncomfortable about adding complexity to the ofdm code just for one particular platform.
1. I'm not that keen on alternate code paths - two pieces of code doing the same thing.  It means more to maintain and test.  I'm flexible on this if there is a good reason for it.  
1. Could this be done with a simple(r) change to `quisk_ccfFilter()` to just output the real part of the signal?
1. Does this break `freedv_comptx()` for 700D?
1. The shape factor of the filter is really up to us, I don't have any firm specs.  I guess stop band  >50dB as IIRC that's a typical Ham spurious spec.  Actually that 50dB is perhaps the product of the Tx sidebands (which might already be attenuated) and filter response. Perhaps before changing the filter coeffs we should establish some requirements.  It's all been a bit ad-hoc until now :thinking: 
1. Re suggested filter coeff changes - what is more important, CPU or shape factor?  Shape factor effectively sets the bandwidth of the on-air signal.  Perhaps we should set some requirements here.
1. However the filter coeffs also affect PAPR and hence performance so good to avoid changing the filter coeffs unless there is a really good reason, and we have checked the PAPR.  We could perhaps consider that step in another PR if there is a good reason for it.
1. Do you have a feel/measurement for the time spent in the filter code with the optimisation?  for example compare `--txbpf 0` to `--txbpf 1` with the optimisations in this PR.  So we can measure the CPU used by the filter alone.
1. What does the tx spectrum look like compared to the original, complex implementation?  Good to check it's still filtering as expected.

An idea I had along these lines (to really minimise the CPU) is use an IIR filter.  This could be very short (5th order).  They have a bit of phase distortion around the transitions but that's probably OK, as we have per-carrier phase estimation.  That might result in a universal solution (one code path, trivial storage for coeffs) that is very efficient.  Lot of work though ... many filters to develop/test and the idea might not work, e.g. IIR filter are short but don't vectorise well.